requirements.txt
fire==0.7.0
google-cloud-aiplatform[adk,agent_engines]==1.88.0
#langchain-community==0.3.20
#crewai-tools==0.38.1
#wikipedia==1.4.0